Revamp signup part and logic
This commit is a revamp of the signup part.
All logics and now handled by the registration state-component, form
property is based on a combineTemplate from all inputs detected in the form
and a global bus error centralize form errors.
For user, errors are more cleverly handled to improve the feedback.