Update marionette version
Prevent the following JS error to occur in some environment.

backbone.marionette.js:3232 Uncaught TypeError: Cannot read property 'radio' of undefined

Error was due to some kind of Backbone version conflicts with Marionnette.
We were not able to detect precisely why the error was occuring in some environments and not in other ones.